backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pypdf==4.0.1
groq>=0.4.0
python-dotenv==1.0.0
